ChatBot-Backend/requirements.txt
flask
langchain
langchain-community
langchain-huggingface
langchain-google-genai
langchain-text-splitters
chromadb
openai
tiktoken
python-docx
pandas
PyMuPDF
openpyxl
python-dotenv
